leapp-inspector: pre-alpha POC version
The commit is expected to be dropped in future. Keeping it for now
just in a case. Currently the script is just library that could be
used e.g. with interactive python to get some data from the leapp db.

Current design could be better from the point of performance, but
regarding the typical size of the leapp db, it's doesn't seem worthy
to mess the code so much.

The tool is not part of this commit yet.