leapp-inspector: refactoring: begin work on minimized SQL communication
It's trade-off. The code is simplified but performance is lowered.
In the current state, all data related to the specified (or last)
leapp execution are loaded permanently and various functions work
with them:
    (+) basically just this one SQL cmd in the script
    (+) probably more simple script
    (-) loading all data from the db (but db_size < 100MB)
    (-) it means that some functions will be little slower
        (but still <1s?)

Additionally add check for the existance of the leapp.db file. Raise
EnvironmentalError when the file doesn't exist.