leapp-inspector: CLI: parsing: Added adjustable parsing of cmdline
It's expected the script will be extended by people on their own
to handle their daily troubles. That's hand in hand with CLI.
New classes and functionality have been introduced to make possible
to adjust CLI in simple way. See SubCommandBaseClass.

The expected base scheme of CLI for Leapp Inspector is:
    leapp-inspector [generic_options] subcmd [subcmd_options]

In this case, everyone cann add (or change..) subcmd and wanted
options.