leapp-inspector: LeappDatabase.get_logs: update the method
Originally the method returned list. However, I realized that the
method should be probably generator instead to reduce processing
of not needed stuff.

The method has been extended for filtering of logs based on the log
level. However, the data of logs are stored in JSON so the decoding
of data is a little bit more expensive but still usable. In case
the filtering is not needed (set the DEBUG lvl for filter) decoding
is skipped and data are yielded immadiately.