leapp-inspector: CLI parsing: add methods for LeappData* creation
Currently there are two top level cmdline options that affects
results of leapp-inspector on the top level:
  --db
  --context

However, these options has to be used by particular subcommands
otherwise they do not have any effect. Right now, extending generic
options by --context, shows the problem with connection
to subcommands (which can be created out of this script/library).
In case we extend the top level options again in future, we do not
want to update each subcommand separately, to prevent possible
issues (that could be discovered really late and could be hidden
to not "careful" user), extend the LeappInspectorCLI class by new
properties to obtain the LeappDatabase or LeappDataPrinter objects.
When used correctly (after the cmdline is parsed) the objects are
created / initialized regarding the top level options correctly.