leapp-inspector: CLI: New subcommand: actors
The subcommand prints various list of actors (exexcuted, productive)
and as well print various data about actors:
 - metadata
 - produced messages
 - executed shell commands
 - logs

It's possible to set the level of verbosity for logs (ERROR -> DEBUG)
to reduce amount of printed data using the --log-level option. Default
is DEBUG to print all.

As well, two formats of the log printing is supported. By default
the logged data is printed as obtained (JSON format, with some
additional metadata). Or the terminal like format can be used when
the --terminal-like option is specified. The content is in that case
decoded and just the stdout/stderr output is printed (all to stdout
during this print).