leapp-inspector: CLI: inspection cmd: work continues...
The commit is temporary - it's expected to be squashed with the
commit introducing the inspection cmd.

Currently the output is extended to provide data about possible
(and real) errors for every actor.
Problems:
  - format needs to be improved
  - some logs could be preeety long - thinking abo
  - errormsg is hard to read. This one should be threated specially
    to ensure that e.g. new-lines are interpreted. E.g. tracebacks
    are hard to read in the current format and would be better
    to have these interpreted
  - currently depends on LeappDatabase.get_errors() method which
    is waiting for "design review".

Note: currently, I can see that some executions of leapp ends in
an actor (in the middle of IPU let's say) and cannot get any error
report. Probably leapp should trap interruptions and log message
before the end of the process.