Fix print of "Executed shell commands" for Py3
Running `leapp-inspector actors ...` command, one of outputs is
the list of executed shell commands. However, the stored log has
different order of fields on Py2 & Py3 (so in time of leapp execution,
no in time of running leapp-inspector). Due to this, the list of
executed shell commands have been empty for leapp executed via Py3
as the content is obtained using a regexp. Update the regexp to
ignore order of fields.

Note: I know this is not ideal implementation at all. Not sure
whether somewhere in the code more dragons are hidden. So far, it
works right now.